Update mupen64plus-input-bot
The existing version of the mupen64plus-input-bot had a memory leak in
one of the JSON calls which caused the emulator to continually grow in
memory usage which would eventually run out of resources on a system.
The new version solves the memory leak, causing minimal memory increases
for the emulator.

Signed-Off-By: Robert Clark <robdclark@outlook.com>